@@ -3,12 +3,12 @@
   "private": true,
   "//type": "module",
   "engines": {
-    "node": "12",
+    "node": ">=14.3",
     "//node": ">=13.2"
   },
   "dependencies": {
-    "firebase-admin": "^9.0.0",
-    "firebase-functions": "^3.9.0"
+    "firebase-admin": "^9.1.0",
+    "firebase-functions": "^3.9.1"
   },
   "//": {
     "engines.node": [
