@@ -67,13 +91,10 @@
     "//rollup-plugin-filesize": "^9.0.2"
   },
   "engines": {
-    "node": ">=13.2"
+    "node": ">=14.3"
   },
   "//comments": {
     "devDependencies": {
-      "@firebase/testing": [
-        "We use it to populate the emulator in local dev mode."
-      ],
       "@vue/compiler-sfc": [
         "This needs to be the same version as 'dependencies.vue' but did not find a way to ",
         "enforce that. Tried '$npm_package_dependencies_vue' but that doesn't work. npm 6.14.4 on macOS."
