@@ -1,9 +1,10 @@
 /*
-* sample.fns/callables.test.js
+* back-end/test-fns/callables.test.js
 */
-import { test, expect, describe, beforeAll, afterAll } from '@jest/globals'
+import { test, expect, describe } from '@jest/globals'
 
-import { fns } from 'firebase-jest-testing'
+//import { fns } from 'firebase-jest-testing'
+import x from '@firebase/testing'
 
 describe ('Cloud Function callables', () => {
   test ('returns a greeting', async () => {
