@@ -1,34 +1,19 @@
 /*
-* rules-test/symbolsC.test.js
+* back-end/test-rules/symbolsC.test.js
 */
 import { strict as assert } from 'assert'
-import * as firebase from '@firebase/testing'
+import { test, expect, describe, beforeAll } from '@jest/globals'
 
-import './tools/jest-matchers'
-
-import { sessionProm } from './tools/guarded-session'
-
-const FieldValue = firebase.firestore.FieldValue;
+import { dbAuth } from 'firebase-jest-testing/firestoreTestingReadOnly';
+import { FieldValue } from 'firebase-jest-testing/firestoreTesting';
 
 const anyDate = new Date();   // a non-server date
 
-import { test, expect, describe, beforeAll } from '@jest/globals'
-
-// Perform extra tests to see the test data isn't changed by other tests (if it is, our guards didn't work!)
-//
-async function HYGIENE( title, doc, f ) {
-  const o = await doc._dump();
-  //console.trace( "HYGIENE: "+ title, o );   // enable for debugging
-  f(o);
-}
-
 let unauth_symbolsC, auth_symbolsC, abc_symbolsC, def_symbolsC;
 
-beforeAll( async () => {
-  const session = await sessionProm();
-
+beforeAll( () => {
   try {
-    const coll = session.collection('projects/1/symbols');
+    const coll = dbAuth.collection('projects/1/symbols');
 
     unauth_symbolsC = coll.as(null);
     auth_symbolsC = coll.as({uid:'_'});
@@ -103,22 +88,13 @@ describe("'/symbols' rules", () => {
   test('members may do changes to an already claimed (by them) symbol', async () => {
     const s2_mod = { size: 999 };
 
-    await HYGIENE( "Before setting to 999", def_symbolsC.doc("2-claimed"), o => {
-      assert( o.size == 50 );
-      assert(o.claimed.by == "def");
-    });
-
     await Promise.all([
       expect( def_symbolsC.doc("2-claimed").update( s2_mod )).toAllow(),     // claimed by him
       expect( abc_symbolsC.doc("2-claimed").update( s2_mod )).toDeny()      // not claimed by them
     ]);
-
-    await HYGIENE( "After setting to 999", def_symbolsC.doc("2-claimed"), o => {
-      assert( o.size == 50 );
-    });
   });
 
-  // BUG: systematically fails #later
+  // tbd. Cannot figure out why this fails. #help
   test.skip('members may revoke a claim', async () => {
     const s2_revoke = { claimed: FieldValue.delete() };
 
@@ -138,21 +114,10 @@ describe("'/symbols' rules", () => {
 
   test('members may delete a symbol claimed to themselves', async () => {
 
-    await HYGIENE( "Before delete", def_symbolsC.doc("2-claimed"), o => {
-      console.debug( "Has:", o );
-      assert( o.size == 50 );
-      assert(o.claimed && o.claimed.by == "def");
-    });
-
     await Promise.all([
       expect( def_symbolsC.doc("2-claimed").delete()).toAllow(),     // claimed by him
       expect( abc_symbolsC.doc("2-claimed").delete()).toDeny()      // not claimed by them
     ]);
-
-    await HYGIENE( "After delete", def_symbolsC.doc("2-claimed"), o => {
-      assert( o.size == 50 );
-      assert(o.claimed && o.claimed.by == "def");
-    });
   });
 
 });
