@@ -4,7 +4,7 @@
 * Write the data in 'docs.js' to the running emulator's Firestore instance.
 */
 import { docs } from './docs.js'
-import { prime } from 'firebase-jest-testing'
+import { prime } from 'firebase-jest-testing/prime'
 
 //console.info("Priming...");
 await prime(docs);
